gesture_engine/requirements.txt
opencv-python>=4.8
mediapipe>=0.10
python-osc>=1.8
python-rtmidi>=1.5
pyyaml>=6.0
